Started motor testing system and added locking
The motors will test themselves,  you DO NOT want to launch a rocket
with broken servos, It can not lock and have the nose cone and parachute
fall out while launching, mabey breaking the parachute. Or it could not
deploy the parachute and you have a dart about to hit someone/something.